inline APIKeyProvider and remove utility file
